requirements.txt
faster-whisper==1.2.1
sounddevice==0.5.5
scipy==1.17.1
keyboard==0.13.5
mouse==0.7.1
requests==2.32.5
numpy==2.3.5
torch==2.5.1+cu121
pystray==0.19.5
Pillow==12.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
